% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{beatAML}
\alias{beatAML}
\title{beatAML}
\format{
A tibble with three variables:
\describe{
\item{\code{inhibitor}}{names of inhibitors}
\item{\code{patient_id}}{anonymous patient ID}
\item{\code{median}}{the median of drug response at corresponding \code{inhibitor} and \code{patient_id}}
}
}
\source{
Tyner, Jeffrey W., et al.(2018), Functional Genomic Landscape of Acute Myeloid Leukaemia. Nature 562 (7728): 526–31.
}
\usage{
beatAML
}
\description{
The Beat AML program publishes an extensive genomic, drug response, and
clinical dataset on acute myeloid leukemia. We represented drug response data from Beat AML program.
}
\keyword{datasets}
