% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{drugComb}
\alias{drugComb}
\title{drugComb}
\format{
A tibble with three variables:
\describe{
\item{\code{inhibitor}}{names of inhibitors}
\item{\code{cell_line}}{cell_line for each \code{inhibitor}}
\item{\code{median}}{the median of drug response at corresponding \code{inhibitor} and \code{cell_line}}
}

For further details, see
\url{https://academic.oup.com/nar/article-abstract/47/W1/W43/5486743}
}
\source{
\url{https://www.biorxiv.org/content/10.1101/2021.03.18.436040v3.full}
}
\usage{
drugComb
}
\description{
DrugComb collect datasets from several major drug combination studies,
covering 437 923 drug combination experiments with 7 423 800 data points
across 93 human cancer cell lines.
}
\keyword{datasets}
