% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{herbIngredient}
\alias{herbIngredient}
\title{herbIngredient}
\format{
A tibble with two variables:
\describe{
\item{\code{herb}}{Scientific Latin names of herbs avialbale in TCMID database}
\item{\code{pubchem_id}}{unique compound id (CID) of ingredients in pubchem database for available \code{herb} in TCMID database}
}
}
\source{
\url{https://www.frontiersin.org/articles/10.3389/fphar.2020.01319/full#h6}
}
\usage{
herbIngredient
}
\description{
The second version of the TCMID database, which included additional ingredient-specific experimental data based on herbal mass spectrometry spectra, was used as the largest dataset in the TCM herb sector. The herb and ingredient associations from this database are provided here.
}
\keyword{datasets}
