% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-cluster.R
\name{plotCluster}
\alias{plotCluster}
\title{Plot the clusters in one projection of the bipartite graph.}
\usage{
plotCluster(graph, cluster, ...)
}
\arguments{
\item{graph}{A igraph graph object.}

\item{cluster}{A igraph cluster object, usually got from \code{\link{findCluster}}}

\item{...}{Pass to \code{\link[networkD3]{forceNetwork}}}
}
\value{
A networkD3 object, the plot of cluster.
}
\description{
This function will visualize the input igraph cluster objects
and igraph graph objects with membership properties. Different memberships
will be represented by dots of different colors.
}
\examples{
\dontrun{
# load the beatAML data and get the incidence matrix
beatAML_data <- NIMAA::beatAML
beatAML_incidence_matrix <- plotInput(beatAML_data)

# extract the sub-matrix
sub_matrices <- extractSubMatrix(
beatAML_incidence_matrix,
shape = c("Square") # the shapes you want to extract
)

# do clustering
cls2 <- findCluster(
sub_matrices$Rectangular_element_max, # the sub-matrix
dim = 2 # set to 2 to use the other part of graph
)


plotCluster(graph=cls2$graph,cluster = cls2$louvain)
}
}
\seealso{
\code{\link[networkD3]{forceNetwork}}
}
