% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{robertson}
\alias{robertson}
\title{robertson}
\format{
A tibble with two variables:
\describe{
\item{\code{plant}}{names of plants}
\item{\code{pollinator}}{names of pollinators for \code{plant}}
}
}
\source{
\url{http://www.ecologia.ib.usp.br/iwdb/html/robertson_1929.html}
}
\usage{
robertson
}
\description{
The author listed 1429 animal species visiting flowers of 456 plant species
that grew in a small area in soutWhwestern Illinois, USA. Marlin and LaBerge
(2001) describe Robertson’s methods.
}
\keyword{datasets}
