% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score-cluster.R
\name{scoreCluster}
\alias{scoreCluster}
\title{Score the clusters in a projected network of the bipartite networks.}
\usage{
scoreCluster(community, graph, distance_matrix)
}
\arguments{
\item{community}{An igraph community object.}

\item{graph}{An igraph graph object.}

\item{distance_matrix}{A matrix, the distance of graph, usually got from
\code{\link{findCluster}}}
}
\value{
A list of various scores.
}
\description{
This function provides additional internal cluster validity measures such as entropy and coverage. The concept of scoring is according to the weight fraction of all intra-cluster edges relative to the total weight of all edges in the graph. This function requiers the community object, igraph object and distance matrix returned by \code{\link{findCluster}} to analyze.
}
\examples{
# load part of the beatAML data
beatAML_data <- NIMAA::beatAML[1:10000,]

# convert to incidence matrix
beatAML_incidence_matrix <- el2IncMatrix(beatAML_data)

# do clustering
cls <- findCluster(beatAML_incidence_matrix,
  dim = 1, method = "infomap", normalization = FALSE,
  rm_weak_edges = TRUE, comparison = FALSE)

# get the scoring result
scoreCluster(community = cls$infomap, graph = cls$graph,
  distance_matrix = cls$distance_matrix)

}
\seealso{
\code{\link[fpc]{cluster.stats}}, \code{\link{findCluster}}
}
