% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-cluster.R
\name{validateCluster}
\alias{validateCluster}
\title{Validate the accuracy of our clustering of the projection comparing to users'
input.}
\usage{
validateCluster(dist_mat, extra_feature, community)
}
\arguments{
\item{dist_mat}{A matrix, the distance of graph, usually got from}

\item{extra_feature}{A dataframe has only one column indicating the
membership of each nodes (rownames).}

\item{community}{An igraph community object.}
}
\value{
A list containing the measured differences between clustering methods
and reference clustering, corrected rand index and Jaccard similarity.
}
\description{
This function will verify the similarity between different
clustering methods and external features (prior knowledge), which is to
measure the results of clustering methods.
}
\examples{
\dontrun{
# load the beatAML data and get the incidence matrix
beatAML_data <- NIMAA::beatAML
beatAML_incidence_matrix <- plotInput(beatAML_data)

# extract the sub-matrix
sub_matrices <- extractSubMatrix(
beatAML_incidence_matrix,
shape = c("Square") # the shapes you want to extract
)

# do clustering
cls <- findCluster(
sub_matrices$Rectangular_element_max, # the sub-matrix
dim = 1
)

validateCluster(dist_mat = cls$distance_matrix,community = cls$leading_eigen)
}
}
