% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-prediction.R
\name{validatePrediction}
\alias{validatePrediction}
\title{Validate and compare edge prediction methods.}
\usage{
validatePrediction(imputation, refer_community, clustering_args)
}
\arguments{
\item{imputation}{A list ro a matrix, the results of different imputation method(s).}

\item{refer_community}{An igraph community object, usually got from \code{\link{findCluster}}.}

\item{clustering_args}{A list indicating the clutering arguments used in \code{\link{findCluster}}, usually got from \code{\link{findCluster}}.}
}
\value{
A list containing the following indicators, Jaccard similarity/ Dice similarity coefficient/ Rand index/ Minkowski(inversed)/ Fowlkes-Mallows index.
}
\description{
This function compares the imputation approaches for predicting edges using the clustering result of a submatrix with non-missing values as a ground truth.  This function performs the same analysis as the \code{findCluster} function on every imputed incidence matrices independently. Then, using different similarity measures, all imputation approaches are compared to each other, revealing how edge prediction methods affects network communities (clusters). The best method should result in a higher degree of similarity (common node membership) to the non-missing submatrix as a benchmark.
}
\examples{
# load part of the beatAML data and get the incidence matrix
beatAML_data <- NIMAA::beatAML[1:10000,]
beatAML_incidence_matrix <- el2IncMatrix(beatAML_data)

# do clustering
cls <- findCluster(
beatAML_incidence_matrix, # the sub-matrix
dim = 1)

# impute
imputations <- imputeMissingValue(beatAML_incidence_matrix)

# validate the imputation
validation_of_imputation <- validateImputation(
imputation = imputations,
refer_community = cls$fast_greedy,
clustering_args = cls$clustering_args
)
}
