% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{beatAML}
\alias{beatAML}
\title{beatAML}
\format{
A tibble with three variables:
\describe{
\item{\code{inhibitor}}{names of inhibitors}
\item{\code{patient_id}}{anonymous patient ID}
\item{\code{median}}{the median of drug response at corresponding \code{inhibitor} and \code{patient_id}}
}
}
\source{
Tyner, Jeffrey W., et al.(2018), Functional Genomic Landscape of Acute Myeloid Leukaemia. Nature 562 (7728): 526–31. \url{http://vizome.org/aml/}
}
\usage{
beatAML
}
\description{
The Beat AML program produced a comprehensive dataset on acute myeloid leukemia (AML) that included genomic data (i.e., whole-exome sequencing and RNA sequencing), ex vivo drug response, and clinical data. We have included only the drug response data from this program in this package.
}
\keyword{datasets}
