% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{drugComb}
\alias{drugComb}
\title{drugComb}
\format{
A tibble with three variables:
\describe{
\item{\code{inhibitor}}{names of inhibitors}
\item{\code{cell_line}}{cell_line for each \code{inhibitor}}
\item{\code{median}}{the median of drug response at corresponding \code{inhibitor} and \code{cell_line}}
}
}
\source{
Zheng, S., Aldahdooh, J., Shadbahr, T., Wang, Y., Aldahdooh, D., Bao, J., Wang, W., Tang, J. (2021). DrugComb update: a more comprehensive drug sensitivity data repository and analysis portal, Nucleic Acids Research, 49(W1), W174–W184.  \url{https://academic.oup.com/nar/article/49/W1/W174/6290546}
}
\usage{
drugComb
}
\description{
DrugComb is a web-based portal for storing and analyzing drug combination screening datasets, containing 739,964 drug combination experiments across 2320 cancer cell lines and 8397 drugs. Single drug screening was extracted and provided here as a subset of drug combination experiments.
}
\keyword{datasets}
