% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{herbIngredient}
\alias{herbIngredient}
\title{herbIngredient}
\format{
A tibble with two variables:
\describe{
\item{\code{herb}}{names of herbs}
\item{\code{pubchem_id}}{unique id of ingredients in pubchemn for \code{herb}}
}
}
\source{
\url{https://www.frontiersin.org/articles/10.3389/fphar.2020.01319/full#h6}
}
\usage{
herbIngredient
}
\description{
This data set used the second version of the TCMID database, as the largest
dataset in TCM herbs field, which contains richer experimental data
originating from ingredient-specific and herbal mass spectrometry spectra.
}
\keyword{datasets}
