% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIMAA-package.R
\docType{package}
\name{NIMAA}
\alias{NIMAA}
\title{NIMAA: A package for Nominal Data Mining Analysis.}
\description{
The NIMAA package provides 12 main functions that build a pipeline for nominal data mining.
}
