% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIMAA-package.R
\docType{package}
\name{NIMAA}
\alias{NIMAA}
\title{NIMAA: A package for Nominal Data Mining Analysis.}
\description{
The NIMAA package contains 13 main functions that together provide a pipeline for nominal data mining. This package also includes four datasets that can be used to perform various nominal data mining analyses.
}
