% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score-cluster.R
\name{scoreCluster}
\alias{scoreCluster}
\title{Score the clusters in one projection of the bipartite graph.}
\usage{
scoreCluster(community, graph, distance_matrix)
}
\arguments{
\item{community}{An igraph community object.}

\item{graph}{An igraph graph object.}

\item{distance_matrix}{A matrix, the distance of graph, usually got from
\code{\link{findCluster}}}
}
\value{
A list of various scores.
}
\description{
This function will use the community object, graph object and
distance matrix to analyze, mainly using the cluster.stats function in the
fpc package, in addition, it also calculates a ‘coverage’ indicator.
}
\examples{
\dontrun{
# load the beatAML data and get the incidence matrix
beatAML_data <- NIMAA::beatAML
beatAML_incidence_matrix <- plotInput(beatAML_data)

# extract the sub-matrix
sub_matrices <- extractSubMatrix(
beatAML_incidence_matrix,
shape = c("Square") # the shapes you want to extract
)

# do clustering
cls2 <- findCluster(
sub_matrices$Rectangular_element_max, # the sub-matrix
dim = 2 # set to 2 to use the other part of graph
)

# get the scoring result
scoring_result <- scoreCluster(
community = cls2$infomap,
graph = cls2$graph,
distance_matrix = cls2$distance_matrix)
}
}
\seealso{
\code{\link[fpc]{cluster.stats}}, \code{\link{findCluster}}
}
